backend/requirements.txt
absl-py==2.4.0
aiofiles==25.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
fastapi==0.135.1
flatbuffers==25.12.19
fonttools==4.62.1
h11==0.16.0
httptools==0.7.1
idna==3.11
ImageIO==2.37.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
lazy-loader==0.5
llvmlite==0.46.0
matplotlib==3.10.8
mediapipe==0.10.32
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
pillow==12.1.1
platformdirs==4.9.4
pooch==1.9.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
PyMatting==1.1.15
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
referencing==0.37.0
rembg==2.0.73
requests==2.32.5
rpds-py==0.30.0
scikit-image==0.26.0
scipy==1.17.1
six==1.17.0
sounddevice==0.5.5
starlette==0.52.1
tifffile==2026.3.3
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
